pyproject.toml
[project]
name = "par-loop"
version = "0.1.0"
description = "PAR Loop — Promise·Attention·Relationship 순환으로 에이전트 동료 품질을 점검하는 엔진"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[project.scripts]
par-loop = "par_loop.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
